Bugfix. infoPanels weren't aligned with their fields.
In z3cform/templates/behaviour.js, rich textareas are replaced with divs and
TinyMCE editor is then activated inside these divs.

In skin/templates/behaviour/behaviour.js, the infoPanels are recreated with
an absolute position set for each via css. This position is determined by the
infoPanel's offset with regards to the label that contains it.

This offset is calculated before the TinyMCE changes are in effect resulting in the labels
being given absolute positions that were off/incorrect.

I have now wrapped all the z3cform field's html in div.z3cFieldContainer and took
the infoPanels out of the labels. This means that we don't have to give them
absolute positions anymore because they are contained in div.z3cFieldContainer
and will therefore always be aligned with the fields.

This approach is also better because if a wysiwig textarea grows dynamically,
the infoPanels below will now also move down with their fields.
This didn't happen before and they stayed static while fields in the form moved
down (again breaking alignment between fields and infoPanels).